services/ventas/requirements.txt
blinker==1.8.2
click==8.1.7
coverage==7.6.1
exceptiongroup==1.2.2
Flask==3.0.3
greenlet==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.22.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.3.2
pytest-cov==5.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.32
tomli==2.0.1
typing_extensions==4.12.2
Werkzeug==3.0.4
gunicorn==21.2.0
flask-cors